Tweaks
- Replaced the Shotgun's random 3-12 pellets with a flat 6 pellets.
- Updated the Weapon base script to use floats for damage instead of ints.
- Updated the SMG script to round the generated bullet damage to an int to keep it at whole numbers.
- Updated the SMG script to snap the generated bullet damage to the nearest multiplication of 0.5 (e.g. 1.5, 2.0, 2.5, etc.).
- Changed the Shotgun's Common damage range from 1-2 to 1.5-2.5.
- Changed the Shotgun's Epic damage range from 5-6 to 4.5-5.5.
- Changed the Shotgun's Legendary damage range from 7-8 to 6-7.
- Increased all the Shotgun's fire rate values by 50.
- Updated the ShotgunInfoPanel's UI to reflect the changes.
- Swapped the Swap and Drop Weapon inputs so Swap Weapon is E and Drop Weapon is Q.

+++ Assets/Scripts/ScriptableObjects/Weapon.cs
@@ -21,7 +21,7 @@ public class Weapon : ScriptableObject
     protected int epicWeight = 10;
     protected int legendaryWeight = 5;
 
-    protected int bulletDamage;
+    protected float bulletDamage;
     [HideInInspector] public int fireRate;
     protected int critChance;
     
@@ -34,11 +34,11 @@ public class Weapon : ScriptableObject
     [SerializeField] protected Sprite legWeaponSprite;
 
     [Header("Bullet Damage")]
-    [SerializeField] protected int comMinBulDmg;
-    [SerializeField] protected int comMaxBulDmg;
-    [SerializeField] protected int rarMinBulDmg, rarMaxBulDmg;
-    [SerializeField] protected int epiMinBulDmg, epiMaxBulDmg;
-    [SerializeField] protected int legMinBulDmg, legMaxBulDmg;
+    [SerializeField] protected float comMinBulDmg;
+    [SerializeField] protected float comMaxBulDmg;
+    [SerializeField] protected float rarMinBulDmg, rarMaxBulDmg;
+    [SerializeField] protected float epiMinBulDmg, epiMaxBulDmg;
+    [SerializeField] protected float legMinBulDmg, legMaxBulDmg;
 
     [Header("Crit Chance")]
     [SerializeField] protected int comMinCritCha;
